Fix all build log errors and add Python root debug.log output
C++ build fixes:
- cmake/CompilerWarnings.cmake: add /FS to fix C1041 parallel PDB writes
- ship_physics.h: use logf() to fix C4244 double→float (C2220)
- lighting.cpp: static_cast<int> for size_t return (C4267)
- model.cpp: static_cast<int> for 5x size_t→int (C4267)
- asteroid_field_renderer.cpp: cast i to int for hash() args (C4267)
- post_processing.cpp: static_cast in loop init (C4267)
- renderer.cpp: static_cast<float> for int health fields (C4244 ×6)
- station_renderer.cpp: static_cast for size_t→uint, remove unused vars
- atlas_title_screen.cpp: suppress unused scale, fix hh usage (C4189)
- entity_picker.cpp: static_cast double→float (C4244)
- star_map.cpp: static_cast size_t→GLsizei/int, int→float (C4267/C4244)
- LocalLLMBackend.cpp: suppress C4996 for getenv safely
- Atlas/cpp_client/include/ui/atlas/atlas_types.h: create missing header (C1083)

Python root logging:
- Shared/Logging/log_utils.py: also write debug.log to repo root
- Scripts/Validate/validate_root.py: add debug.log to allowlist
- .gitignore: ignore /debug.log (local debug artifact)

Agent-Logs-Url: https://github.com/shifty81/MasterRepoRefactor/sessions/ba642b88-f42d-4d34-89d4-3efc68091380

+++ NovaForge/Client/App/include/core/ship_physics.h
@@ -28,7 +28,7 @@ public:
         
         // Align time: time to reach 75% max velocity
         float getAlignTime() const {
-            return -log(0.25) * getAgility() / 1000000.0f;
+            return -logf(0.25f) * getAgility() / 1000000.0f;
         }
     };
